Move utilities to a separate module; start on using Specialize trait.
