+++ src/dev_utilities/specialize.rs
@@ -1,0 +1,8 @@
+pub trait Specialize<T> : Sized {
+    fn can_specialize(&self) -> bool {
+        false
+    }
+    fn specialize(self) -> Option<T> {
+        None
+    }
+}
